Add ability to use file-based or in-memory db
Now uses $schema->deploy to create the SQLite DB, rather than a
hand-written external script.

Can now switch between using file-based or in-memory DB, based on the
value of $ENV{SQLITE_FILE_DB}. The advantage of a file-based DB is that
it can be accessed from an external connection, which is handy for
debugging tests.