Added new column to sample, expanded taxonomy table
Updated the model to add a new "patient_location" column to the "sample"
table. Re-generated the DBIC modules and changed tests and test data
files accordingly.

Expanded the "taxonomy" look-up table with the extra columns necessary
for storing the taxonomy tree.

Removed the ResultSet for the "antimicrobial_resistance" table, since it
was empty anyway.