Add ontology loading and re-factor ontology checking
The ontology terms were previously checked on loading using
relationships in the database. Those relationships also prevent us
reloading the ontologies in the future, since to do so would break
foreign key relationships.

Instead of letting the DB do the checking, we'll add checking in the
code and let DBIC enforce the relationship when trying to load samples.