Tax and ontology loaders record file details
Added code to the schema to add records to the external_resources table
and added calls to that methods to the loading scripts for the taxonomy
tree and the ontologies.

The idea is that there is a record of when a particular external
resource was loaded from file. This will only work if the resource is
loaded through the relevant load_* script, but it's better than nothing.