Update some columns with more sensible types
Initial choices were somewhat arbitrary, so most column types will need
to be tweaked to make them more sensible. The ID columns on the ontology
tables are good examples of column types that don't allow any room to
grow, if the ontologies add extra digits to their IDs, for instance.