Fix a failing test due to module versions
The test was developed using MooseX::Params::Validate v0.19, under which
it passed. When running the same test with MX::P::V v0.21, a different
error message is generated (same meaning, different details), and the
test fails.

Updated the module on the dev box and fixed the test to pass on the
newer version.