Fix tests that were broken by DB changes
The schema was changed to put a unique constraint on the sample
accession column, which necessitated changing the tests that didn't
follow that rule. Changed the tests themselves, the data files that they
use, and the test DB schema definition.

The only changes to code in the this commit are due to the removal of
commented blocks.