Update deleted_at when adding duplicate samples
When a new sample is added, we check if it already exists and, if so,
update the existing row, setting its "deleted_at" field to the current
time. The new row is then added and becomes the "live" row.

Updated methods dealing with samples to take notice of "deleted_at" when
searching.