Set up mechanism for "deleting" rows
We don't actually want to delete anything, so we'll flag "deleted" rows
by setting their "deleted_at" field to the time/date of deletion.
Similarly we don't want to update rows, but simply replace them with new
values.

This commit removes the "updated_at" column from the tables and adds a
role ("Undeletable") that handles deletion by setting "deleted_at" on
the deleted rows. Some tables have sub-tables that need to have rows
deleted too, so we're doing that cascade here too.

Updated the schema build script to apply the role automatically to the
Result classes when they're built.