Fix failures if environment variables set in shell
The module relies on environment variables, such as HICF_SCRIPT_CONFIG,
which are set to various values by the test script. If the variables are
already set in the shell that's running the tests, the first tests fail,
because they expect the environment variable to be unset.

Fix the problem by explicitly removing the required variables from the
%ENV hash.