Move read/write components out of main schema
The tables in the main schema will be accessed primarily through a
read-only account via the website. The user and midas_session tables
need to be accessed using a read/write account, since the website needs
to be able to update user data and store session data.

We need to split the database in two, creating one database for the HICF
data and another for the user-related data.

This commit includes a new script, bin/build_schemata.pl, which
generates the two sets of DBIC wrappers. The table definitions for
user and midas_session are also moved into the new namespace, along with
the Role that is applied to User.