Only load the first 45 chars of a term description
The live database runs in STRICT_TRANS_TABLES mode, meaning that MySQL
will throw an error if we attempt to insert a value larger than the
column allows. (The test database has no such restriction...)

We don't care about the ontology descriptions, because the ontology
tables are used purely as a look up for the term ID. This commit simply
truncates the "name" value to 45 characters before it's inserted into
the database.