Remove relationship between sample and location
The location table stores the latitude and longitude for the sample
locations specified by the GAZ ontology term. Since we'll only geocode
those terms when they're actually used, when we load sample data it's
likely that we won't have values for lat/long at that point.

Adding a "might_have" relationship between sample and location seemed
like a good idea, but it causes problems when testing because the SQLite
test DB is instantiated from the DBIC schema definition. That processes
adds a foreign key to the sample table, which isn't found in the live
MySQL instance. Finally, loading ontologies involves deleting the
contents of the gazetteer table, which runs into the foreign key
constraint and causes test failures.

Rather than try to work around this now, we'll remove the relationship
and re-add it if we actually need it later.