Re-add commented check for unknown tax ID/sci name
Re-adding the statements that allow tax ID and scientific name to be
unknown, but adding them as commented text. Although this check is
probably bad for the HICF use case, it might be important if the code is
re-used and it makes sense to have that particular check in place.