Retro-fit newer methods to understand deleted samples
This commit fixes two newer methods that retrieve all or some samples,
so that they now accept a boolean flag to tell them whether to retrieve
samples which have been deleted. Rows in the sample table can be flagged
as deleted by setting their "deleted_at" date, which is what happens if
a newer version of the metadata is loaded.