Fix broken test
The schema method that returns summary information appears to work fine
with the live MySQL database but was causing problems with the SQLite
test database. This commit changes the method so that it doesn't try to
chain queries together like it did, which seems to fix the issue. Also
had to tweak the associated tests.