Fix AMR RS and add_multiple_midas_users
In script, make sure that the database has returned a passphrase
before mailing the user with their account details. Also, skip comment
lines in the input file (lines starting with "#").

In AMR ResultSet, there's a regex that checks the error message that
comes back from DBIC and makes it a bit more human-readable. Turns out
the error message from DBIC is capitalised when running on the mac, but
not on linux, presumable because of version drift between the two sets
of DBIC / DBI drivers. This commit makes the regex case insensitive
which makes the test pass on both platforms.