Fix manifest count in summary
The summary gives the total number of manifests in the database, but it
previously simply counted the number of rows in the manifest table. That
could be misleading, since a manifest can be reloaded and all of its
samples flagged as deleted. In that case the first manifest really
shouldn't be counted, if we want the number in the summary to reflect
the number of manifests that have contributed live data, rather than
simply the number that we've loaded.

This commit adds a more complicated query to determine the number of
manifests, which looks also at the number of live samples. Added a
test to check it.