Fix query that counts SIR results for summary
The "get_sample_summary" method previously counted the number of AMR
*results* with a given susceptibility, rather than the number of samples
having AMR results with the given susceptbility. This commit fixes that.

Also, the HICF checklist will be updated to add a new SIR flag, "U"
(unknown), so that code is factored into the new code. Updated the tests
to account for that.